+++ internal/test/mocks/collections_store.go
@@ -12,13 +12,13 @@ type GetCollectionsFunc func(ctx context.Context, userID int32, limit int, offse
 
 type GetCollectionFunc func(ctx context.Context, userID int32, nodeID string) (collections.GetCollectionResponse, error)
 
-type DeleteCollectionFunc func(ctx context.Context, collectionID int64) error
+type DeleteCollectionFunc func(ctx context.Context, collectionID int32) error
 
-type UpdateCollectionFunc func(ctx context.Context, userID int32, collectionID int64, update collections.UpdateCollectionRequest) (collections.GetCollectionResponse, error)
+type UpdateCollectionFunc func(ctx context.Context, userID int32, collectionID int32, update collections.UpdateCollectionRequest) (collections.GetCollectionResponse, error)
 
-type StartPublishFunc func(ctx context.Context, collectionID int64, userID int32, publishingType publishing.Type) error
+type StartPublishFunc func(ctx context.Context, collectionID int32, userID int32, publishingType publishing.Type) error
 
-type FinishPublishFunc func(ctx context.Context, collectionID int64, publishingStatus publishing.Status, strict bool) error
+type FinishPublishFunc func(ctx context.Context, collectionID int32, publishingStatus publishing.Status, strict bool) error
 
 type CollectionsStore struct {
 	CreateCollectionsFunc
@@ -90,28 +90,28 @@ func (c *CollectionsStore) GetCollection(ctx context.Context, userID int32, node
 	return c.GetCollectionFunc(ctx, userID, nodeID)
 }
 
-func (c *CollectionsStore) DeleteCollection(ctx context.Context, collectionID int64) error {
+func (c *CollectionsStore) DeleteCollection(ctx context.Context, collectionID int32) error {
 	if c.DeleteCollectionFunc == nil {
 		panic("mock DeleteCollection function not set")
 	}
 	return c.DeleteCollectionFunc(ctx, collectionID)
 }
 
-func (c *CollectionsStore) UpdateCollection(ctx context.Context, userID int32, collectionID int64, update collections.UpdateCollectionRequest) (collections.GetCollectionResponse, error) {
+func (c *CollectionsStore) UpdateCollection(ctx context.Context, userID int32, collectionID int32, update collections.UpdateCollectionRequest) (collections.GetCollectionResponse, error) {
 	if c.UpdateCollectionFunc == nil {
 		panic("mock UpdateCollection function not set")
 	}
 	return c.UpdateCollectionFunc(ctx, userID, collectionID, update)
 }
 
-func (c *CollectionsStore) StartPublish(ctx context.Context, collectionID int64, userID int32, publishingType publishing.Type) error {
+func (c *CollectionsStore) StartPublish(ctx context.Context, collectionID int32, userID int32, publishingType publishing.Type) error {
 	if c.StartPublishFunc == nil {
 		panic("mock StartPublish function not set")
 	}
 	return c.StartPublishFunc(ctx, collectionID, userID, publishingType)
 }
 
-func (c *CollectionsStore) FinishPublish(ctx context.Context, collectionID int64, publishingStatus publishing.Status, strict bool) error {
+func (c *CollectionsStore) FinishPublish(ctx context.Context, collectionID int32, publishingStatus publishing.Status, strict bool) error {
 	if c.FinishPublishFunc == nil {
 		panic("mock FinishPublish function not set")
 	}
